OutlineDraft
Introduction
Many people shop daily, in fact, studies show that approximately 40 million people shop on the weekend alone but only 34 percent of Americans budget and actually maintain it. It is difficult creating a budget let alone maintaining one when there are so many deals, discounts, and coupons. People tend to forget that even though an item may be for a discounted rate and once that money is spent regardless if 10 dollars was saved, it’s gone!
Although there are many budgeting or money management apps on the market, they don’t seem to be decreasing consumers spending habits and increasing their money consciousness. In this project, the Save & Shop budgeting application will help consumers to become more aware of their spending habits while decreasing regretful overspending. This can be done by allowing consumers to: (1) manage their budget by tracking their spending, (2) create as many budgets as desired with allocated max monthly spending, and (3) transactions can be added and applied to the budget.








1.	REQUIREMENTS ANALYSYSIS AND SYSTEM SPECIFICATIONS

1.1	Software Requirements Specification Document
1.1.1	Data Requirements
The set of data that is involved in any project is defined using data requirements. For this project, the main data required is the date and the amount of money. Without this the application cannot process the money spent.
1.1.2	Functional Requirements
Functional requirements are properties are properties that must exist in the final system. Mobile applications need to be downloaded from the play store. Applications can be free to download and other require payment before download. This application, Save & Shop, is free. To access this application and all its features, the user needs to open after installing. 
1.1.3	Performance Requirements
The application must have rapid response time, tolerance, platform dependencies, scalability, and user friendly. The app should respond rapidly when in use, it should be scalable so that in future new features are readily accepted and should be able to deliver error messages in case of network issues.



2.	SYSTEM DESIGN

2.1	Design Approach 
This application was designed in a way to make a seamless working application with all integrated parts working together.
2.2	User Interface Design
The user interface should always be simple for a more enjoyable experience for the user. Save & Shop has only a few navigational clicks to avoid the user being frustrated because they cannot navigate the application. 
3.	IMPLEMENTATION


3.1	Introduction to Programming Languages, IDE’S, Tools and Technologies used for the Implementation
3.1.1	Java
Since the project is about developing an Android Application, the default programming language is Java. All Android applications are made using Java in the Android Studio or Eclipse. 
3.1.2	IDE’s, Tools and Technologies:
3.1.2.1	Android Studio

Android Studio is designed for developing Android applications. It consists of all Android SDK tools to design, develop, maintain, test, debug and publish our app. The IDE is designed very efficiently which makes the developer’s job much easier. It also supports Git as a version control system to maintain the app changes and push them into GitHub. All java files, layout files (for design) are integrated into a single project easily. After the completion of project, the whole application could be put as an Android Package) file, in which we run that APK file in any device and use the application.

3.1.3.2 Android Software Development Kit:
One of the main tools used in developing android applications, as it packages many core features into one SDK, and it can be used in the application easily. The helps decrease the amount of code being written and building the application faster.

3.1.3.3 SDK Manager
A tool that maintains the updates of the already installed software for the project to run correctly. SDK Manager also notifies us when the project is not compatible with devices.